+++ packages/interactivity/src/index.ts
@@ -69,6 +69,7 @@ export { useState, useRef } from 'preact/hooks';
  * ```
  */
 export const watch = effect;
+export { renderElement } from './render';
 
 const requiredConsent =
 	'I acknowledge that using private APIs means my theme or plugin will inevitably break in the next version of WordPress.';